requirements.txt
Django==2.2.17
django-cpf==0.1.0
django-filter==2.4.0
django-phone-field==1.8.1
django-stdimage==5.1.1
djangorestframework==3.12.2
importlib-metadata==3.1.1
Markdown==3.3.3
Pillow==8.0.1
progressbar2==3.53.1
python-utils==2.4.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
zipp==3.4.0
